add repo added
