+++ src/index.ts
@@ -1,32 +1,12 @@
 #!/usr/bin/env node
-import { AxiosError } from 'axios';
 import { promises as fs } from 'fs';
-import {
-	LanguageId,
-	LogType,
-	MetaCallJSON
-} from 'metacall-protocol/deployment';
-import {
-	generateJsonsFromFiles,
-	generatePackage,
-	PackageError
-} from 'metacall-protocol/package';
-import API from 'metacall-protocol/protocol';
-import { join } from 'path';
+
 import args from './cli/args';
-import { input } from './cli/inputs';
 import { ins } from './cli/inspect';
-import { apiError, error, info, printLanguage, warn } from './cli/messages';
-import Progress from './cli/progress';
-import {
-	containerSelection,
-	fileSelection,
-	languageSelection,
-	planSelection
-} from './cli/selection';
-import { logs } from './logs';
+import { error } from './cli/messages';
+import { planSelection } from './cli/selection';
+import { deployFromRepository, deployPackage } from './deploy';
 import { startup } from './startup';
-import { zip } from './utils';
 
 enum ErrorCode {
 	Ok = 0,
@@ -37,187 +17,48 @@ enum ErrorCode {
 }
 
 void (async () => {
-	const rootPath = args['workdir'];
-	const name = args['projectName'].toLowerCase();
 	const inspect = args['inspect'];
 
 	if (inspect) await ins();
 
-	try {
-		if (!(await fs.stat(rootPath)).isDirectory()) {
-			error(`Invalid root path, ${rootPath} is not a directory.`);
-			return process.exit(ErrorCode.NotDirectoryRootPath);
-		}
-	} catch (e) {
-		error(`Invalid root path, ${rootPath} not found.`);
-		return process.exit(ErrorCode.NotFoundRootPath);
+	if (!args['workdir'] && !args['addrepo']) {
+		error('Either provide work directory or repository url to deploy');
 	}
 
-	try {
-		const config = await startup();
-		let descriptor = await generatePackage(rootPath);
+	// TODO: We should cache the plan and ask for it only once
+	const plan =
+		args['plan'] ||
+		(await planSelection('Please select plan from the list'));
 
-		const deploy = async (additionalJsons: MetaCallJSON[]) => {
-			// TODO: We should cache the plan and ask for it only once
-			const plan =
-				args['plan'] ||
-				(await planSelection('Please select plan from the list'));
+	const config = await startup();
 
-			const api = API(config.token as string, config.baseURL);
-			const enabled = await api.deployEnabled();
+	//if workdir is passed call than deploy using package
+	if (args['workdir']) {
+		const rootPath = args['workdir'];
 
-			if (!enabled) {
-				error('Your account is not enabled to deploy');
-				return process.exit(ErrorCode.AccountDisabled);
+		try {
+			if (!(await fs.stat(rootPath)).isDirectory()) {
+				error(`Invalid root path, ${rootPath} is not a directory.`);
+				return process.exit(ErrorCode.NotDirectoryRootPath);
 			}
-
-			const { progress, pulse, hide } = Progress();
-
-			const archive = await zip(
-				rootPath,
-				descriptor.files,
-				progress,
-				pulse,
-				hide
-			);
-
-			// TODO: We should do something with the return value, for example
-			// check for error or show the output to the user
-			await api.upload(
-				name,
-				archive,
-				additionalJsons,
-				descriptor.runners
-			);
-
-			// TODO: We can ask for environment variables here too and cache them
-			/*
-			const { enableEnv } = await prompt<{ enableEnv: boolean }>([
-				{
-					type: 'confirm',
-					name: 'enableEnv',
-					message: 'Add env vars?',
-					default: false
-				}
-			]);
-			const env = enableEnv
-				? await prompt<{ env: string }>([
-						{
-							type: 'input',
-							name: 'env',
-							message: 'Type env vars in the format: K1=V1, K2=V2'
-						}
-				]).then(({ env }) =>
-						env
-							.split(',')
-							.map(kv => {
-								const [k, v] = kv.trim().split('=');
-								return { [k]: v };
-							})
-							.reduce((obj, kv) => Object.assign(obj, kv), {})
-				)
-				: {};
-			*/
-
-			info(`Deploying ${rootPath}...\n`);
-
-			try {
-				await api.deploy(name, [], plan);
-			} catch (err) {
-				apiError(err as AxiosError);
-			}
-
-			// TODO: Need a TUI for logs
-
-			try {
-				const container: string = await containerSelection([
-					...descriptor.runners,
-					'deploy'
-				]);
-				const type =
-					container === 'deploy' ? LogType.Deploy : LogType.Job;
-
-				await logs(container, name, type);
-			} catch (err) {
-				apiError(err as AxiosError);
-			}
-		};
-
-		const createJsonAndDeploy = async (saveConsent: string) => {
-			const potentialPackages = generateJsonsFromFiles(descriptor.files);
-			const potentialLanguages = Array.from(
-				new Set<LanguageId>(
-					potentialPackages.reduce<LanguageId[]>(
-						(langs, pkg) => [...langs, pkg.language_id],
-						[]
-					)
-				)
-			);
-
-			const languages = await languageSelection(potentialLanguages);
-			const packages = potentialPackages.filter(pkg =>
-				languages.includes(pkg.language_id)
-			);
-
-			for (const pkg of packages) {
-				pkg.scripts = await fileSelection(
-					`Select files to load with ${printLanguage(
-						pkg.language_id
-					)}`,
-					pkg.scripts
-				);
-			}
-
-			const cacheJsons = saveConsent === 'Y' || saveConsent === 'YES';
-
-			const additionalPackages = cacheJsons
-				? await (async () => {
-						for (const pkg of packages) {
-							await fs.writeFile(
-								join(
-									rootPath,
-									`metacall-${pkg.language_id}.json`
-								),
-								JSON.stringify(pkg, null, 2)
-							);
-						}
-
-						// If they are cached, genearte the descriptor again
-						descriptor = await generatePackage(rootPath);
-
-						// The descriptor already contains the packages so
-						// there is no need to send additional packages
-						return [];
-				  })()
-				: packages; // Otherwise, packages are not cached, send them
-
-			await deploy(additionalPackages);
-		};
-
-		switch (descriptor.error) {
-			case PackageError.None: {
-				await deploy([]);
-				break;
-			}
-			case PackageError.Empty: {
-				error(`The directory you specified (${rootPath}) is empty`);
-				return process.exit(ErrorCode.EmptyRootPath);
-			}
-			case PackageError.JsonNotFound: {
-				warn(
-					`No metacall.json was found in ${rootPath}, launching the wizard`
-				);
-
-				const askToCachePackagesFile = (): Promise<string> =>
-					input('Do you want to save metacall.json file? (Y/N): ');
-
-				await createJsonAndDeploy(
-					(await askToCachePackagesFile()).toUpperCase()
-				);
-				break;
-			}
+		} catch (e) {
+			error(`Invalid root path, ${rootPath} not found.`);
+			return process.exit(ErrorCode.NotFoundRootPath);
 		}
-	} catch (e) {
-		console.error(e);
+
+		try {
+			await deployPackage(config, plan);
+		} catch (e) {
+			console.error(e);
+		}
+	}
+
+	//if addrepo is passed then deploy from repository url
+	if (args['addrepo']) {
+		try {
+			await deployFromRepository(config, plan);
+		} catch (e) {
+			console.error(e);
+		}
 	}
 })();
